feat(core): remove platform modules, move event-manager services to KitEventManagerModule
Closes #167

BREAKING CHANGE: KitPlatformBrowserModule, KitPlatformBrowserModule are not available now
