--- packages/core/src/kit-platform-browser/README.md
@@ -1,3 +1,0 @@
-# Platform-browser
-
-Browser implementations.
