--- packages/core/src/kit-platform-browser/kit-platform-browser.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { KitEventManagerService } from '../kit-event-manager/kit-event-manager.service';
-import { KitBrowserEventManagerService } from './event-manager/kit-browser-event-manager.service';
-
-@NgModule({
-  providers: [
-    {
-      provide: KitEventManagerService,
-      useClass: KitBrowserEventManagerService,
-    },
-  ],
-})
-export class KitPlatformBrowserModule {
-}
