--- packages/core/src/kit-platform-browser/kit-platform-browser.ts
@@ -1,1 +1,0 @@
-export * from './kit-platform-browser.module';
