--- packages/core/src/kit-platform-server/README.md
@@ -1,3 +1,0 @@
-# Platform-server
-
-Server rendering implementations.
