--- packages/core/src/kit-platform-server/kit-platform-server.ts
@@ -1,1 +1,0 @@
-export * from './kit-platform-server.module';
