+++ playground/collection/app/app.module.ts
@@ -3,7 +3,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterModule } from '@angular/router';
-import { KitModule, KitPlatformBrowserModule, KitRootModule } from '@ngx-kit/core';
+import { KitModule, KitRootModule } from '@ngx-kit/core';
 import { CollectionDemoModule } from '../../../packages/collection/demo/collection-demo.module';
 import { AppComponent } from './app.component';
 import { PlaygroundComponent } from './playground/playground.component';
@@ -21,7 +21,6 @@ import { PlaygroundComponent } from './playground/playground.component';
     RouterModule.forRoot([]),
     KitRootModule,
     KitModule,
-    KitPlatformBrowserModule,
     CollectionDemoModule,
   ],
   providers: [],
