+++ playground/core/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { KitModule, KitPlatformBrowserModule, KitRootModule } from '@ngx-kit/core';
+import { KitModule, KitRootModule } from '@ngx-kit/core';
 import { AppComponent } from './app.component';
 import { VcComponent } from './vc/vc.component';
 
@@ -12,7 +12,6 @@ import { VcComponent } from './vc/vc.component';
     BrowserAnimationsModule,
     ReactiveFormsModule,
     FormsModule,
-    KitPlatformBrowserModule,
     KitRootModule,
     KitModule,
   ],
