+++ website/app/app-server.module.ts
@@ -2,7 +2,6 @@ import { HttpClient } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { ServerModule } from '@angular/platform-server';
 import { ModuleMapLoaderModule } from '@nguniversal/module-map-ngfactory-loader';
-import { KitPlatformServerModule } from '@ngx-kit/core';
 import { AppModule } from './app.module';
 import { RootComponent } from './root/root.component';
 import { ServerHttp } from './server/server-http';
@@ -11,7 +10,6 @@ import { ServerHttp } from './server/server-http';
   imports: [
     AppModule,
     ServerModule,
-    KitPlatformServerModule,
     ModuleMapLoaderModule,
   ],
   bootstrap: [
